Fixed dot generation using sin/cos functions

+++ src/Dot.java
@@ -1,11 +1,11 @@
 public class Dot{
     private int num;
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     public Dot(int newNum){ num = newNum; }
 
-    public Dot(int newNum, int newX, int newY){
+    public Dot(int newNum, double newX, double newY){
         num = newNum;
         x = newX;
         y = newY;
@@ -13,11 +13,11 @@ public class Dot{
 
     public int getNum() { return num; }
 
-    public void setX(int x) { this.x = x; }
+    public void setX(double x) { this.x = x; }
 
-    public int getX(){ return x; }
+    public double getX(){ return x; }
 
-    public void setY(int y) { this.y = y; }
+    public void setY(double y) { this.y = y; }
 
-    public int getY(){ return y; }
+    public double getY(){ return y; }
 }
